Ready for 0.2.1 release
